refactor: flatten SelectField internals hierarchy
