+++ src/SelectField/Select.js
@@ -3,13 +3,13 @@ import React, { Component, Fragment, createRef } from 'react'
 import css from 'styled-jsx/css'
 import cx from 'classnames'
 
-import { colors } from '../../theme'
+import { colors } from '../theme'
 import {
     innerSpacingSides,
     inputFontSizeValue,
     inputKinds,
     inputSizes,
-} from '../../forms/constants'
+} from '../forms/constants'
 
 const styles = css`
     select {
