+++ src/SelectField/index.js
@@ -3,10 +3,10 @@ import React, { Fragment } from 'react'
 import css from 'styled-jsx/css'
 import cx from 'classnames'
 
-import { Arrow } from './SelectField/Arrow'
+import { Arrow } from './Arrow'
 import { LabelFilled } from '../forms/LabelFilled'
 import { LabelOutlined } from '../forms/LabelOutlined'
-import { Select } from './SelectField/Select'
+import { Select } from './Select'
 import { colors } from '../theme'
 import { iconStatusPropType, iconStatuses } from '../icons/constants'
 import { inputKinds, inputSizes } from '../forms/constants'
